Moviendose a BigDecimal

+++ src/test/java/dds/monedero/model/MonederoTest.java
@@ -7,6 +7,8 @@ import dds.monedero.exceptions.SaldoMenorException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class MonederoTest {
@@ -19,50 +21,50 @@ public class MonederoTest {
 
   @Test
   void Poner() {
-    cuenta.poner(1500);
+    cuenta.poner(BigDecimal.valueOf(1500));
   }
 
   @Test
   void PonerMontoNegativo() {
-    assertThrows(MontoNegativoException.class, () -> cuenta.poner(-1500));
+    assertThrows(MontoNegativoException.class, () -> cuenta.poner(BigDecimal.valueOf(-1500)));
   }
 
   @Test
   void TresDepositos() {
-    cuenta.poner(1500);
-    cuenta.poner(456);
-    cuenta.poner(1900);
+    cuenta.poner(BigDecimal.valueOf(1500));
+    cuenta.poner(BigDecimal.valueOf(456));
+    cuenta.poner(BigDecimal.valueOf(1900));
   }
 
   @Test
   void MasDeTresDepositos() {
     assertThrows(MaximaCantidadDepositosException.class, () -> {
-          cuenta.poner(1500);
-          cuenta.poner(456);
-          cuenta.poner(1900);
-          cuenta.poner(245);
+          cuenta.poner(BigDecimal.valueOf(1500));
+          cuenta.poner(BigDecimal.valueOf(456));
+          cuenta.poner(BigDecimal.valueOf(1900));
+          cuenta.poner(BigDecimal.valueOf(245));
     });
   }
 
   @Test
   void ExtraerMasQueElSaldo() {
     assertThrows(SaldoMenorException.class, () -> {
-          cuenta.setSaldo(90);
-          cuenta.sacar(1001);
+          cuenta.setSaldo(BigDecimal.valueOf(90));
+          cuenta.sacar(BigDecimal.valueOf(1001));
     });
   }
 
   @Test
   public void ExtraerMasDe1000() {
     assertThrows(MaximoExtraccionDiarioException.class, () -> {
-      cuenta.setSaldo(5000);
-      cuenta.sacar(1001);
+      cuenta.setSaldo(BigDecimal.valueOf(5000));
+      cuenta.sacar(BigDecimal.valueOf(1001));
     });
   }
 
   @Test
   public void ExtraerMontoNegativo() {
-    assertThrows(MontoNegativoException.class, () -> cuenta.sacar(-500));
+    assertThrows(MontoNegativoException.class, () -> cuenta.sacar(BigDecimal.valueOf(-500)));
   }
 
 }
